+++ ios/Sources/PamNative/Protocol/PamProtocol.swift
@@ -407,6 +407,7 @@ public enum PamConstants {
     public static let workletIterations = 444
     public static let navigationBarHidden = 445
     public static let borderStyle = 446
+    public static let scrollTargetAlignment = 447
     public static let sectionItems = 80
     public static let hostProperties = 100
     public static let opacity = 38